restore: support sparse restores also on windows
